Criado DTO para receber os dados via rest e implementado o mvc do lançamento

+++ src/main/java/br/com/goldhirsch/repository/LancamentoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.goldhirsch.model.Lancamento;
 
 @Repository
-public interface LancamentoRepository extends JpaRepository<Lancamento, Long> {
+public interface LancamentoRepository extends JpaRepository<Lancamento, Integer> {
 
 }
